game: compare NudgeCursor's bounds as doubles and widen stage unsigned
NudgeCursor converts each edge with __floatsidf and compares it through
__cmpdf2 against doubles at 0x007f0288 to 0x007f02a0, and
LevelStats::Load zero-extends the stage byte with lbu at 0x00142b2c, so
the bounds are now double constants and the stage local is unsigned
char.

+++ src/game/freqappearancedetail.cpp
@@ -44,8 +44,8 @@ constexpr unsigned kMeshCopyFlags = 0;
 // tenth of them, inside these bounds.
 constexpr double kHalf = 0.5;
 constexpr float kEdgeInset = 0.1f;
-constexpr int kLowBound = -55;
-constexpr int kHighBound = 55;
+constexpr double kLowBound = -55.0;
+constexpr double kHighBound = 55.0;
 
 // A part of this category cannot rise above the zero line, and randomize() does not swap its
 // template. randomize() does not swap the templates of the other two categories.
